Fix graph creation for logWhatIChange
The cycle removal needs to be delayed until after we know if the graph
needs to be reversed, otherwise it will only be possible to go
from two nodes in a two way dependency relation in one direction when it
should be possible to go both ways.

In other works, 'getData' was returning an empty object when the graph
was reversed because the original graph did not have the back arrows.